Initial Files
Uploaded all of the UI interface I've dont till date. I learnt a lot making this website. I learnt how to use Figma for the designing and in the process learnt quite a large amount of information. Hopefully by the time i upload the next set of files. I'll have staretd with the backend and finsihed most of the UI pages required.